add graph for baseline usage rate
This rate includes a value that is is derived from the uploaded historical energy use for the same period last year. A weekday and weekend average use is calculated for each hour of the bill period (e.g., energy use for weekdays 2:00 - 3:00 p.m., weekends 12:00 - 1:00 p.m., etc.) based on this historical data.